pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-tg-bot"
version = "0.1.0"
description = "Telegram bridge for local CLI agent workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0,<2.0",
  "python-telegram-bot>=21,<22",
]

[project.scripts]
llm-tg-bot = "llm_tg_bot.main:main"

[tool.setuptools]
packages = ["llm_tg_bot"]
